Funsiona, falta tema de minimapa grande

+++ include/cub3d.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cub3d.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aarenas- <aarenas-@student.42malaga.com    +#+  +:+       +#+        */
+/*   By: jcallejo <jcallejo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/30 13:31:24 by aarenas-          #+#    #+#             */
-/*   Updated: 2025/02/19 18:32:13 by aarenas-         ###   ########.fr       */
+/*   Updated: 2025/02/20 11:06:02 by jcallejo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -120,7 +120,7 @@ typedef struct s_data
 	int				map_rows;
 	int				map_longest_row;
 	int				mousex;
-	int				animcounter;
+	float			animcounter;
 	float			tx;
 	float			ty;
 	float			ty_step;
